patches now store the actual 6901 pointer in the patch file
i removed the concept of successful patches (since they are now all part of the history, and can be queried by identity).

it makes it so that the Patch object follows 6902 closer, and also saves ~2% on disk space, and also "makes more sense as a url". this means that i now have the responsibility to manage some escaping but it is not complex