BREAKING: moved match schedule to a separate location
this prevents editing the schedule from being destructive to match recording data. this strikes a better balance between normalization and ease of access to data. This also was able to show off some bugs in the UI for edge cases and behaviors that were previously inconsistent or undefined, specifically with what data shows where and alliance colors